+++ skrub/_apply_to_cols.py
@@ -129,6 +129,22 @@ class SingleColumnTransformer(BaseEstimator):
                 wrapped = _wrap_add_check_single_column(getattr(subclass, method))
                 setattr(subclass, method, wrapped)
 
+    def get_feature_names_out(self, input_features=None):
+        """Get the output feature names.
+
+        Parameters
+        -----------
+        input_features : array-like of str, default=None
+            Input feature names. Ignored.
+
+        Returns
+        --------
+        all_outputs_
+            The names of the output features.
+        """
+        check_is_fitted(self, "all_outputs_")
+        return self.all_outputs_
+
 
 def _wrap_add_check_single_column(f):
     # as we have only a few predefined functions to handle, using their exact